Add changes

+++ kylieq-CIS415F19-P2/part3.c
@@ -46,8 +46,6 @@ void SigChldHandler(int sig_num) {
                 printf("Process: %d - Terminated\n", PCBS[COUNTER]->pid);
                 
             }
-            else if(status > 0)
-                PCBS[COUNTER]->exit_status = 1;
         }
 }
 
@@ -65,7 +63,7 @@ void SigAlrmHandler(int sig_num) {
     }
     else {
         while(1) {
-            if (PCBS[COUNTER]->state == RUNNING) {
+            if (PCBS[COUNTER]->state == RUNNING && PCBS[COUNTER]->exit_status != 1) {
                 kill(PCBS[COUNTER]->pid, SIGSTOP);
                 printf("Process: %d - Received Signal SIGALRM - Suspended\n", PCBS[COUNTER]->pid);
                 PCBS[COUNTER]->state = STOPPED;
@@ -91,7 +89,7 @@ void SigAlrmHandler(int sig_num) {
         }
 
         while(1) {
-            if (PCBS[COUNTER]->state == STOPPED) {
+            if (PCBS[COUNTER]->state == STOPPED && PCBS[COUNTER]->exit_status != 1) {
                 kill(PCBS[COUNTER]->pid, SIGCONT);
                 printf("Process: %d - Received Signal SIGALRM - Continued\n", PCBS[COUNTER]->pid);
                 PCBS[COUNTER]->state = RUNNING;
